==> 01-ExplorationData.ipynb <==
import pandas as pd
import numpy as np
import json
import datetime as dt
import seaborn as sns
import matplotlib.pyplot as plt

with open('   ', 'r') as file:
    data = json.load(file)
    
df = pd.json_normalize(data)
df.head(1)


# Variables
df.columns

# Checking missing values
df.isnull().sum()

# Checking missing values in the Revenue variable:
df[['revenue.2020','revenue.2019','revenue.2018','revenue.2017','revenue.2016','revenue.2015']].isnull().sum()
# Dropping missing values Revenue variable
df= df.dropna(subset=['revenue.2020','revenue.2019','revenue.2018','revenue.2017','revenue.2016', 'revenue.2015'], how='all')


### Checking variables

nulls = df[['province','company_category']].isnull().sum()
nulls

df.province.value_counts()

# Recoding 'province' variable 
def code_province(x):
    if x =='Antwerp':
        return 1
    if x =='Vlaams Brabant':
        return 2
    if x =='East-Flanders':
        return 3
    if x =='West-Flanders':
        return 4
    if x =='Limburg':
        return 5

df['province_code'] = df['province'].apply(code_province)


df.company_category.value_counts()

# Recoding 'company size' variable
def code_size(x):
    if x =='Small':
        return 1
    if x =='Medium sized':
        return 2
    if x =='Large':
        return 3
    if x =='Very large':
        return 4

df['size_code'] = df['company_category'].apply(code_size)


#Transform string to date
df['DateFormat'] = pd.to_datetime(df.creation_date, format="%Y/%m/%d")
#Extracting Year
df['creation_year'] = df['DateFormat'].dt.year

df.legal_form.value_counts()

# Recoding 'legal_form' variable
def code_form(x):
    if x =='Private company with limited liability':
        return 1
    if x =='Private company':
        return 2

df['legal_form_code'] = df['legal_form'].apply(code_form)

### Correlation Matrices

corrMatrix = df[['revenue.2020','ebit.2019','ebitda.2019','profit_and_loss_after_taxes.2019','total_assets.2019','financial_profit_and_loss.2019','staff_count.2019','net_added_value.2019','staff_costs.2019']].corr()
sns.heatmap(corrMatrix, annot=True)
plt.show()

# Plot, we can see outliers:
fig = plt.figure(figsize=(9,5))
sns.scatterplot(x='province', y='revenue.2020', data=df)
#plt.title('Quality vs fixed acidity', fontsize = 15, fontweight = 'bold')
#plt.xlabel('Quality', fontsize=10)
#plt.ylabel('Fixed acidity', fontsize=10)

# Plot revenue 2020
fig = plt.figure(figsize=(9,5))
sns.barplot(x='province', y='revenue.2020', data=df)
plt.title('revenue.2020', fontsize = 15, fontweight = 'bold')
plt.xlabel('province', fontsize=10)
plt.ylabel('revenue.2020', fontsize=10)

#Plot revenue 2019 vs province
fig = plt.figure(figsize=(9,5))
sns.barplot(x='province', y='revenue.2019', data=df)
plt.title('revenue 2019', fontsize = 12, fontweight = 'bold')
plt.xlabel('province', fontsize=10)
plt.ylabel('revenue 2019', fontsize=10)

# Correlations
corrMatrix = df[['size_code','revenue.2020','revenue.2019']].corr()
sns.heatmap(corrMatrix, annot=True)
plt.show()

# Correlations
corrMatrix = df[['staff_costs.2019','staff_count.2019','revenue.2020']].corr()
sns.heatmap(corrMatrix, annot=True)
plt.title('Correlation')
plt.show()

# Correlations
corrMatrix = df[['revenue.2020','staff_costs.2019','staff_count.2019','revenue.2019']].corr()
sns.heatmap(corrMatrix, annot=True)
plt.title('Correlation')
plt.show()

# Correlations
corrMatrix = df[['revenue.2020', 'size_code', 'ebit.2019', 'ebitda.2019','total_assets.2019','staff_costs.2019','revenue_transf.2019','revenue.2019']].corr()
sns.heatmap(corrMatrix, annot=True)
plt.title('Correlation')
plt.show()

# Correlations
corrMatrix = df[['revenue.2020', 'size_code', 'ebit.2019', 'ebitda.2019','total_assets.2019','staff_costs.2019','revenue_transf.2019','revenue.2019', 'profit_and_loss_after_taxes.2020']].corr()
sns.heatmap(corrMatrix, annot=True)
plt.title('Correlation')
plt.show()

# Correlations
corrMatrix = df[['revenue.2020', 'size_code', 'ebitda.2019','total_assets.2019', 'financial_profit_and_loss.2019','staff_costs.2019','revenue_transf.2019','revenue.2019']].corr()
sns.heatmap(corrMatrix, annot=True)
plt.title('Correlation')
plt.show()
