==> 03-Models_evaluation.ipynb <==
# import libraries
import pandas as pd
import numpy as np

# Sklearn processing
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler

# Sklearn regression algorithms
from sklearn.neighbors import KNeighborsRegressor
from sklearn.tree import DecisionTreeRegressor
from sklearn.svm import SVR
from xgboost import XGBRegressor
from sklearn.linear_model import Ridge
from sklearn.linear_model import Lasso
from sklearn.linear_model import ElasticNet
from sklearn.ensemble import RandomForestRegressor

# Sklearn regression model evaluation function
from sklearn.metrics import mean_absolute_error
from sklearn.metrics import mean_squared_error
from sklearn.metrics import r2_score
from sklearn.metrics import explained_variance_score

# Importing the dataset
df = pd.read_csv(" ")

# drop the unwanted columns
df.drop(['Unnamed: 0','vat_number','nace_code','highLevel_NaceCode',
                'province','year'], axis='columns', inplace=True)

# Select Features and Target
X = df.iloc[:, :-1].values
y = df.iloc[:, -1].values

# Split data into Test and Training Sets
test_size = 0.33
seed = 1
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = test_size, random_state = seed)

#Rescale data to avoid dominate lower value features by higher value features when calculating distances
trans = StandardScaler()
X_train = trans.fit_transform(X_train)
X_test = trans.transform(X_test)

Evaluating different models

#Create a list of all the Models to fit and check
models = [KNeighborsRegressor(), DecisionTreeRegressor(), SVR(), XGBRegressor(), Ridge(), 
              Lasso(), ElasticNet(), RandomForestRegressor()]

# loop through all the models, fit them and check for different metrics
for model in models:
    model.fit(X_train, y_train)
    y_pred = model.predict(X_test)
    print(type(model).__name__, mean_absolute_error(y_test, y_pred))
    print(type(model).__name__, r2_score(y_test, y_pred))
    print(type(model).__name__, explained_variance_score(y_test, y_pred))
    print(type(model).__name__, mean_squared_error(y_test, y_pred))
